ai4se/requirements.txt
transformers>=4.36.0
torch>=2.1.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.16.0
optimum>=1.16.0
onnx>=1.15.0
onnxruntime>=1.16.0  # Faster CPU inference during ONNX export
huggingface-hub>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Amharic Voice Support
openai-whisper>=20231117  # Speech-to-Text for Amharic
piper-tts>=1.2.0  # Text-to-Speech for Amharic
soundfile>=0.12.0  # Audio file I/O
